Refactor greeting logic and update corresponding tests
Refactored the `Main` method in `Program.cs` to call a new method
`WelcomeUser()` for writing the welcome message to the console.
Added the `WelcomeUser` method to improve code organization by
separating concerns. Updated the test method in `ConsoleTests.cs`
to reflect this change, renaming it from `ShouldWriteHelloWorld`
to `ShouldWelcomeUser` and updating the assertion to expect
"Welcome User!\r\n" instead of "Hello, World!\r\n".