Add GetStateCode method and corresponding unit test
Added a new method `GetStateCode` in the `Program` class within the `DotnetStarter.Logic.ElephantCarpaccio` namespace. This method prompts the user to enter a state code and returns the input as a string.

Also added a new test method `ShouldGetStateCodeAsInputFromUser` in the `ConsoleTests` namespace within the `Consoletests.cs` file. This test method sets up a `StringReader` to simulate user input of "CA", calls the `Program.GetStateCode` method, and asserts that the returned value is "CA".